+++ Firmware/tmc2130.h
@@ -11,7 +11,7 @@ extern uint8_t tmc2130_current_r[4];
 extern uint8_t tmc2130_mres[4];
 
 //flags for axis stall detection
-extern uint8_t tmc2130_sg_thr[4];
+extern int16_t tmc2130_sg_thr[4];
 
 extern bool tmc2130_sg_stop_on_crash;
 extern uint8_t tmc2130_sg_crash; //crash mask
